refactor: fix app structure
