+++ youtube-client/src/app/app.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import HeaderComponent from './header/header/header.component';
-import SearchResultsComponent from './search/search-results/search-results.component';
+import HeaderComponent from './core/components/header/header.component';
+import SearchResultsComponent from './youtube/components/search-results/search-results.component';
 
 @Component({
   selector: 'app-root',
