+++ youtube-client/src/app/core/components/search-bar/search-bar.component.scss
