+++ youtube-client/src/app/core/components/search-overlay/search-overlay.component.scss
