+++ youtube-client/src/app/youtube/directives/colored-border.directive.ts
