+++ youtube-client/src/app/youtube/pipes/sort-pipe.pipe.ts
@@ -1,6 +1,6 @@
 /* eslint-disable class-methods-use-this */
 import { Pipe, PipeTransform } from '@angular/core';
-import Item from '../search/search-item.model';
+import Item from '../../youtube/models/search-item.model';
 
 @Pipe({
   name: 'sortPipe',
